--- test/utils/user-event/focus.js
@@ -1,12 +1,0 @@
-import { getActiveElement, isFocusable, eventWrapper } from './utils';
-
-function focus(element) {
-  if (!isFocusable(element)) return;
-
-  const isAlreadyActive = getActiveElement(element.ownerDocument) === element;
-  if (isAlreadyActive) return;
-
-  eventWrapper(() => element.focus());
-}
-
-export default focus;
